+++ src/main/java/com/laakarisimu/view/ISimulaattorinUI.java
@@ -14,9 +14,10 @@ public interface ISimulaattorinUI {
 	public void setLaakarinPalvelemat(int laakarinPalvelematAs);
 	public void setLaakarinPalkka(double palkka);
 	public void setSairaanhoitajanPalkka(double palkka);
-	public void setHoidontarveLieva(int lieva);
-	public void setHoidontarveKohtalainen(int kohtalainen);
-	public void setHoidontarveVakava(int vakava);
+	public void setHoidontarveLieva(double lieva);
+	public void setHoidontarveKohtalainen(double kohtalainen);
+	public void setHoidontarveVakava(double vakava);
+
 	
 	//Kontrolleri antaa käyttöliittymälle tuloksia, joita Moottori tuottaa 
 	public void setLoppuaika(double aika);
